This commits contains some tweeks and a working frontend for viewing journeys. It uses Astro and Svelte to function. I also added some configurations. There is still work to be done considering pagination (most likely) and viewing all stations/single stations

+++ flyway/sql/V1___initial_schema.sql
@@ -4,6 +4,6 @@ CREATE TABLE journey (
     DepartureStation TEXT NOT NULL,
     ReturnID INT NOT NULL,
     ReturnStation TEXT NOT NULL,
-    Distance INT NOT NULL,
-    Duration INT NOT NULL
+    Distance FLOAT NOT NULL,
+    Duration FLOAT NOT NULL
 );
